moved compression tests into their own file
